Moving tile out of lingo home into tile folder
