+++ client/src/components/browse/lingo-home.js
@@ -1,38 +1,6 @@
 import React from 'react';
 
-function pastelColors(){
-    var darkFactor = 0.75
-    var r = parseInt((Math.round(Math.random()* 127) + 127) * darkFactor ).toString(16);
-    var g = parseInt((Math.round(Math.random()* 127) + 127) * darkFactor ).toString(16);
-    var b = parseInt((Math.round(Math.random()* 127) + 127) * darkFactor ).toString(16);
-    return '#' + r + g + b;
-}
-
-var LingoTile = React.createClass({
-
-        getInitialState: function() {
-
-            return {
-                bgColor: pastelColors(),
-                active: false,
-                onClick: (function(){
-                    this.setState( { 'active' : !this.state.active } );
-                }).bind( this )
-            };
-        },
-        render: function() {
-
-            var classString = 'tile col-xs-12 col-sm-6 col-md-4';
-
-            return (
-                <div onClick={this.props.onClick.bind( null, this.props.term.name )} 
-                	className={classString} 
-                	style={{background:this.state.bgColor}}>
-                    <h2 className="centered">{this.props.term.name}</h2>
-                </div>
-            );
-        }
-});
+import LingoTile from '../tile/Tile'; 
  
 //table for the whole thing
 var LingoTable = React.createClass({
@@ -48,7 +16,7 @@ var LingoTable = React.createClass({
             <div className="container">
             
                 <div className="row">{
-                    this.props.terms.map(term => <LingoTile term={term} key={term.id} onClick={this.setActive} />)
+                    this.props.terms.map(term => <LingoTile term={term.name} key={term.id} onClick={this.setActive} />)
                 }</div>
              </div>
         );
